requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==1.0.0
    # via
    #   tensorboard
    #   tensorflow
astunparse==1.6.3
    # via tensorflow
cachetools==5.0.0
    # via google-auth
certifi==2021.10.8
    # via
    #   httplib2shim
    #   requests
charset-normalizer==2.0.11
    # via requests
earthengine-api==0.1.296
    # via -r requirements.in
flatbuffers==2.0
    # via tensorflow
future==0.18.2
    # via earthengine-api
gast==0.4.0
    # via tensorflow
google-api-core==2.4.0
    # via
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-storage
google-api-python-client==1.12.10
    # via earthengine-api
google-auth==2.5.0
    # via
    #   earthengine-api
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
    #   tensorboard
google-auth-httplib2==0.1.0
    # via
    #   earthengine-api
    #   google-api-python-client
google-auth-oauthlib==0.4.6
    # via tensorboard
google-cloud-core==2.2.2
    # via google-cloud-storage
google-cloud-storage==2.1.0
    # via earthengine-api
google-crc32c==1.3.0
    # via google-resumable-media
google-pasta==0.2.0
    # via tensorflow
google-resumable-media==2.1.0
    # via google-cloud-storage
googleapis-common-protos==1.54.0
    # via google-api-core
grpcio==1.43.0
    # via
    #   tensorboard
    #   tensorflow
h5py==3.6.0
    # via tensorflow
httplib2==0.20.2
    # via
    #   earthengine-api
    #   google-api-python-client
    #   google-auth-httplib2
    #   httplib2shim
httplib2shim==0.0.3
    # via earthengine-api
idna==3.3
    # via requests
importlib-metadata==4.10.1
    # via markdown
keras==2.7.0
    # via tensorflow
keras-preprocessing==1.1.2
    # via tensorflow
libclang==13.0.0
    # via tensorflow
markdown==3.3.6
    # via tensorboard
numpy==1.22.1
    # via
    #   h5py
    #   keras-preprocessing
    #   opt-einsum
    #   tensorboard
    #   tensorflow
oauthlib==3.2.0
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
protobuf==3.19.4
    # via
    #   google-api-core
    #   google-cloud-storage
    #   googleapis-common-protos
    #   tensorboard
    #   tensorflow
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.7
    # via httplib2
requests==2.27.1
    # via
    #   google-api-core
    #   google-cloud-storage
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.8
    # via google-auth
six==1.16.0
    # via
    #   absl-py
    #   astunparse
    #   earthengine-api
    #   google-api-python-client
    #   google-auth
    #   google-auth-httplib2
    #   google-pasta
    #   grpcio
    #   httplib2shim
    #   keras-preprocessing
    #   tensorflow
tensorboard==2.8.0
    # via tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow==2.7.0
    # via -r requirements.in
tensorflow-estimator==2.7.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.23.1
    # via tensorflow
termcolor==1.1.0
    # via tensorflow
typing-extensions==4.0.1
    # via tensorflow
uritemplate==3.0.1
    # via google-api-python-client
urllib3==1.26.8
    # via
    #   httplib2shim
    #   requests
werkzeug==2.0.2
    # via tensorboard
wheel==0.37.1
    # via
    #   astunparse
    #   tensorboard
    #   tensorflow
wrapt==1.13.3
    # via tensorflow
zipp==3.7.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
